learning resource api changes
