+++ generics/services/sunbird.js
@@ -8,450 +8,35 @@
 //dependencies
 
 const request = require('request');
-const shikshalokamService = require(ROOT_PATH+"/generics/helpers/shikshalokam");
 const fs = require("fs");
 
-/**
-  * Generate Dial codes
-  * @function
-  * @name generateCodes
-  * @param dialCodeData - body data for generating dial code.
-  * @param dialCodeData.count - Count of dial code required.
-  * @param dialCodeData.publisher - Publisher name.
-  * @param token - Logged in user token.
-  * @returns {Promise}
-*/
-
-var generateCodes = async function ( dialCodeData,token ) {
-
-    const generateDialCodeUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_GENERATE_DIALCODE;
-
-    return new Promise(async (resolve,reject)=>{
-        
-        let options = {
-            "headers":{
-            "content-type": "application/json",
-            "authorization" :  process.env.AUTHORIZATION,
-            "x-authenticated-user-token" : token,
-            "x-channel-id" : process.env.SUNBIRD_ORGANISATION_ID 
-            },
-            json : dialCodeData
-        };
-        
-        request.post(generateDialCodeUrl,options,callback);
-        
-        function callback(err,data){
-            if( err ) {
-                return reject({
-                    message : constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                });
-            } else {
-                let dialCodeData = data.body;
-                return resolve(dialCodeData);
-            }
-        }
-    })
-    
-}
-
-/**
-  * Publish dial code
-  * @function
-  * @name publishCode
-  * @param dialCodeData - body data for generating dial code.
-  * @param codeId - publish code based on unique id.
-  * @param token - logged in user token .
-  * @param dialCodeData
-  * @returns {Promise}
-*/
-
-var publishCode = async function ( codeId,token,dialCodeData ) {
-
-    const publishDialCodeUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_PUBLISH_DIALCODE+"/"+codeId;
-
-    return new Promise(async (resolve,reject)=>{
-        
-        let options = {
-            "headers":{
-            "content-type": "application/json",
-            "authorization" : process.env.AUTHORIZATION,
-            "x-authenticated-user-token" : token,
-            "x-channel-id" : process.env.SUNBIRD_ORGANISATION_ID 
-            },
-            json : dialCodeData
-        };
-
-        request.post(publishDialCodeUrl,options,callback);
-            
-        function callback(err,data){
-            if( err ) {
-                return reject({
-                    message : constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                });
-            } else {
-                let dialCodeData = data.body;
-                return resolve(dialCodeData);
-            }
-        }
-    })
-}
-
-/**
-  * Get the status of the published code.
-  * @function
-  * @name codeStatus
-  * @param dialCodeData
-  * @param token - logged in user token .
-  * @returns {String} status code
-*/
-
-var codeStatus = async function ( token,codeData ) {
-
-    const dialCodeStatusUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_DIALCODE_STATUS;
-
-    return new Promise(async (resolve,reject)=>{
-        
-        let options = {
-            "headers":{
-            "content-type": "application/json",
-            "authorization" :  process.env.AUTHORIZATION,
-            "x-authenticated-user-token" : token,
-            "x-channel-id" : process.env.SUNBIRD_ORGANISATION_ID 
-            },
-            json : codeData
-        };
-
-        request.post(dialCodeStatusUrl,options,callback);
-            
-        function callback(err,data){
-            if( err ) {
-                return reject({
-                    message : constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                });
-            } else {
-                let dialCodeStatus = data.body;
-                return resolve(dialCodeStatus);
-            }
-        }
-    })
-    
-}
-
-/**
-  * Link the content data.
-  * @function
-  * @name linkContent
-  * @param contentData - Link content data.
-  * @param token - logged in user token.
-  * @returns {String} status code
-*/
-
-var linkContent = async function ( token,contentData ) {
-
-    const linkContentUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_CONTENT_LINK;
-
-    return new Promise(async (resolve,reject)=>{
-        
-        let options = {
-            "headers":{
-            "content-type": "application/json",
-            "authorization" :  process.env.AUTHORIZATION,
-            "x-authenticated-user-token" : token,
-            "x-channel-id" : process.env.SUNBIRD_ORGANISATION_ID 
-            },
-            json : contentData
-        };
-
-        request.post(linkContentUrl,options,callback);
-            
-        function callback(err,data){
-            if( err ) {
-                return reject({
-                    message : constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                })
-            } else {
-                let linkContentData = data.body;
-                return resolve(linkContentData);
-            }
-        }
-    })
-    
-}
-
-/**
-  * Publish content.
-  * @function
-  * @name publishContent
-  * @param contentData - Content data.
-  * @param contentId - Publish content id.
-  * @returns {String}
-*/
-
-var publishContent = async function ( contentData,contentId ) {
-
-    const publishContentUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_PUBLISH_CONTENT+"/"+contentId;
-
-    return new Promise(async (resolve,reject)=>{
-        try {
-
-            if( !global.publisherToken ) {
-
-                global.publisherToken = 
-                await shikshalokamService.generateKeyCloakAccessToken(
-                    process.env.SUNBIRD_PUBLISHER_USERNAME,
-                    process.env.SUNBIRD_PUBLISHER_PASSWORD 
-                );   
-            }
-            
-            let options = {
-                "headers":{
-                    "content-type": "application/json",
-                    "authorization" : process.env.AUTHORIZATION,
-                    "x-authenticated-user-token" :  global.publisherToken.token ,
-                    "x-channel-id" : process.env.SUNBIRD_ORGANISATION_ID 
-                },
-                json : contentData
-            };
-
-            request.post(publishContentUrl,options,callback);
-            
-            function callback(err,data){
-                if( err ) {
-                    throw {
-                        message : 
-                        constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                    };
-                } else {
-                    let publishContentData = data.body;
-                    return resolve(publishContentData)
-                }
-            }
-        } catch(err) {
-            return reject(err);
-        }
-    })
-    
-}
-
-
-/**
-  * Get user profile information.
-  * @function
-  * @name getUserProfile
-  * @param token - Logged in user token.
-  * @param userId - Logged in user id.
-  * @returns {JSON} - user profile information
-*/
-
-var getUserProfile = async function ( token,userId ) {
-
-    const userProfileUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_USER_READ+"/"+userId;
-
-    return new Promise(async (resolve,reject)=>{
-        try {
-            
-            let options = {
-                "headers": {
-                    "content-type" : "application/json",
-                    "authorization" : process.env.AUTHORIZATION,
-                    "x-authenticated-user-token" :  token
-                }
-            };
-
-            request.get(userProfileUrl,options,callback);
-            
-            function callback(err,data) {
-                if( err) {
-                    throw {
-                        message : 
-                        constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                    };
-                } else {
-                    if(data.statusCode != 200) {
-                        return resolve({
-                            responseCode : "SUNBIRD_SERVICE_ERROR"
-                        })
-                    } else {
-                        let userProfileInformationData = data.body;
-                        return resolve(JSON.parse(userProfileInformationData))
-                    }
-                }
-            }
-        } catch(err) {
-            return reject(err);
-        }
-    })
-    
-}
-
-/**
-  * Sync index
-  * @function
-  * @name indexSync
-  * @param syncData - Sync 
-  * @param syncData.requests.objectType
-  * @param syncData.requests.userIds {Array} - user ids
-  * @param token - Logged in user token.
-  * @returns {Promise}
-*/
-
-var indexSync = async function ( syncData,token ) {
-
-    const indexSyncUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_INDEX_SYNC;
-
-    return new Promise(async (resolve,reject)=>{
-        
-        let options = {
-            "headers":{
-            "content-type": "application/json",
-            "authorization" :  process.env.AUTHORIZATION,
-            "x-authenticated-user-token" : token,
-            "x-channel-id" : process.env.SUNBIRD_ORGANISATION_ID 
-            },
-            json : syncData
-        };
-        
-        request.post(indexSyncUrl,options,callback);
-        
-        function callback(err,data){
-            if( err ) {
-                return reject({
-                    message : constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                });
-            } else {
-                let indexSyncedData = data.body;
-                return resolve(indexSyncedData);
-            }
-        }
-    })
-    
-}
-
-/**
-  * Create content data
-  * @function
-  * @name createContent
-  * @param contentData - content data.
-  * @param token - Logged in user token.
-  * @returns {Promise}
-*/
-
-var createContent = async function ( contentData,token ) {
-
-    const contentUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_CREATE_CONTENT;
-
-    return new Promise(async (resolve,reject)=>{
-        
-        const options = {
-            "headers" : {
-                "content-type": "application/json",
-                "authorization" : process.env.AUTHORIZATION,
-                "x-authenticated-user-token" : token,
-                "x-channel-id" : process.env.SUNBIRD_ORGANISATION_ID 
-            },
-            json : contentData
-        };
-        
-        request.post(contentUrl,options,callback);
-        
-        function callback(err,data){
-            if( err ) {
-                return reject({
-                    message : constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                });
-            } else {
-                let contentData = data.body;
-                return resolve(contentData);
-            }
-        }
-    })
-    
-}
 
 /**
   * Lists of organisations
   * @function
   * @name organisationList
-  * @param requestData - requested data
+  * @param token - user keyclock token
+  * @param limit - page limit
+  * @param offset - page offset
   * @returns {Promise}
 */
 
-var organisationList = async function (requestData) {
+var organisationList = async function (token,limit,offset) {
 
-    const organisationUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_ORGANISATION_LISTS;
+    const organisationUrl =process.env.SUNBIRD_SERIVCE_HOST +
+    process.env.SUNBIRD_SERIVCE_BASE_URL +
+    process.env.URL_PREFIX + constants.endpoints.SUNBIRD_ORGANISATION_LISTS+"?limit="+limit+"&page="+offset;
 
     return new Promise(async (resolve,reject)=>{
         
         let options = {
-            "headers" : {
-                "content-type": "application/json",
-                "authorization" :  process.env.AUTHORIZATION
-            },
-            json : requestData
-        };
-        
-        request.post(organisationUrl,options,callback);
-        
-        function callback(err,data){
-            if( err ) {
-                return reject({
-                    message : constants.apiResponses.SUNBIRD_SERVICE_DOWN
-                });
-            } else {
-                let contentData = data.body;
-                return resolve(contentData);
-            }
-        }
-    })
-    
-}
-
-/**
-  * Upload content data
-  * @function
-  * @name uploadContent
-  * @param file - Upload file data.
-  * @param contentId - content id.
-  * @param token - logged in token.
-  * @param contentType - content type.
-  * @param mimeType - mime type.
-  * @returns {Promise}
-*/
-
-var uploadContent = async function ( file,contentId,token,contentType,mimeType ) {
-
-    const contentUrl = 
-    process.env.sunbird_url+constants.endpoints.SUNBIRD_UPLOAD_CONTENT + `/${contentId}`;
-
-    return new Promise(async (resolve,reject)=>{
-        
-        const options = {
-            "headers" : {
-                "content-type" : contentType,
+            "headers": {
+                "content-type" : "application/json",
                 "authorization" : process.env.AUTHORIZATION,
-                "x-authenticated-user-token" : token,
-                "x-channel-id" : process.env.SUNBIRD_ORGANISATION_ID 
-            },
-            formData : {
-                "fileName" : {
-                    "value" : fs.createReadStream(file),
-                    "options" : {
-                        contentType : mimeType
-                    }
-                }
+                "x-authenticated-user-token" :  token
             }
         };
-        
-        request.post(contentUrl,options,callback);
+        request.post(organisationUrl,options,callback);
         
         function callback(err,data){
             if( err ) {
@@ -466,6 +51,8 @@ var uploadContent = async function ( file,contentId,token,contentType,mimeType )
     })
     
 }
+
+
 
 /**
   * Get user profile.
@@ -499,7 +86,6 @@ var userProfile = async function ( userId,token ) {
         
         function callback( err,data ) {
             if( err ) {
-                console.log(err)
                 return reject({
                     message : constants.apiResponses.SUNBIRD_SERVICE_DOWN
                 });
@@ -509,19 +95,126 @@ var userProfile = async function ( userId,token ) {
             }
         }
     })
-    
 }
 
+/**
+  * Call to sunbird service. 
+  * @function
+  * @name callToSunbird
+  * @param requestBody - Logged in user Id.
+  * @param token - Logged in user token.
+  * @param url - url of the api call.
+  * @param requestType - http request method
+  * @returns {JSON} - sunbird service response
+*/
+
+function callToSunbird(requestType, url, token="", requestBody = "") {
+    return new Promise(async (resolve, reject) => {
+
+        let options = {
+            "headers": {
+                "content-type": "application/json",
+                "internal-access-token":process.env.INTERNAL_ACCESS_TOKEN
+            }
+        };
+        if(token){
+            options['headers']["x-authenticated-user-token"] = token;
+        }  
+
+        if (requestType != "GET") {
+            options['json'] = requestBody;
+        }
+
+        url = process.env.SUNBIRD_SERIVCE_HOST + process.env.SUNBIRD_SERIVCE_BASE_URL + process.env.URL_PREFIX + url;
+        if (requestType == "PATCH") {
+            request.patch(url, options, callback);
+        } else if (requestType == "GET") {
+            request.get(url, options, callback);
+        } else {
+            request.post(url, options, callback);
+        }
+
+        function callback(err, data) {
+
+            if (err) {
+                 return reject({
+                    message: constants.apiResponses.SUNBIRD_SERVICE_DOWN
+                });
+            } else {
+                return resolve(data.body);
+            }
+        }
+
+    });
+}
+
+/**
+* To get list of learning resources
+* @method
+* @name  learningResources
+* @param {String} token - user access token.
+* @param {String} pageSize - page size of the request
+* @param {String} pageNo - page no of the request
+* @param {String} category - category for the learning resource
+* @param {String} subCategory - subcategory for the learning resource
+* @param {String} topic -  topic for the learning resource
+* @param {String} language - language for the learning resource
+* @param {String} sortBy - sortBy option for the learning resource
+* @returns {json} Response consists of list of learning resources
+*/
+
+const learningResources = function (token, pageSize, pageNo, filters, sortBy,searchText) {
+    return new Promise(async (resolve, reject) => {
+        try {
+
+            let learningResourceApiUrl = constants.endpoints.SUNBIRD_LEARNING_RESOURCE_LIST;
+            learningResourceApiUrl = learningResourceApiUrl + "?limit=" + pageSize + "&page=" + pageNo;
+            if(searchText){
+                learningResourceApiUrl = learningResourceApiUrl + "&search="+searchText;
+            }
+            if(sortBy){
+                learningResourceApiUrl = learningResourceApiUrl + "&sortBy=" + sortBy;
+            }
+         
+            let mappedFilterList = {};
+            let filterKeys = Object.keys(filters);
+            
+            if (filterKeys && filterKeys.length > 0) {
+                filterKeys.map(filter => {
+                    let mappingType = "";
+                   
+                    if (filter == "category") {
+                        mappingType = "board"
+                    } else if (filter == "subCategory") {
+                        mappingType = "gradeLevel"
+                    } else if (filter == "topic") {
+                        mappingType = "medium"
+                    } else if (filter == "language") {
+                        mappingType = "subject"
+                    }else if(filter == "mimeType"){
+                        mappingType = "mimeType"
+                    }
+                    mappedFilterList[mappingType] = filters[filter];
+                });
+            }
+
+            let requestBody = {
+                filters: mappedFilterList
+            }
+
+            let response = await callToSunbird("POST", learningResourceApiUrl, token, requestBody);
+            return resolve(response);
+        } catch (error) {
+            reject({ message: constants.apiResponses.SUNBIRD_SERVICE_DOWN });
+        }
+
+
+    })
+}
+
+
 module.exports = {
-    generateCodes : generateCodes,
-    publishCode : publishCode,
-    codeStatus : codeStatus,
-    linkContent : linkContent,
-    publishContent : publishContent,
-    getUserProfile : getUserProfile,
-    indexSync : indexSync,
-    createContent : createContent,
-    uploadContent : uploadContent,
     organisationList : organisationList,
-    userProfile : userProfile
+    userProfile : userProfile,
+    learningResources: learningResources
 };
